+++ tests/Rule/SymbolRuleTest.php
@@ -9,21 +9,21 @@ use PHPUnit\Framework\TestCase;
 use Stadly\PasswordPolice\ValidationError;
 
 /**
- * @coversDefaultClass \Stadly\PasswordPolice\Rule\Symbol
+ * @coversDefaultClass \Stadly\PasswordPolice\Rule\SymbolRule
  * @covers ::<protected>
  * @covers ::<private>
  */
-final class SymbolTest extends TestCase
+final class SymbolRuleTest extends TestCase
 {
     /**
      * @covers ::__construct
      */
     public function testCanConstructRuleWithMinConstraint(): void
     {
-        $rule = new Symbol('$%&@!', 5, null);
+        $rule = new SymbolRule('$%&@!', 5, null);
 
         // Force generation of code coverage
-        $ruleConstruct = new Symbol('$%&@!', 5, null);
+        $ruleConstruct = new SymbolRule('$%&@!', 5, null);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -32,10 +32,10 @@ final class SymbolTest extends TestCase
      */
     public function testCanConstructRuleWithMaxConstraint(): void
     {
-        $rule = new Symbol('$%&@!', 0, 10);
+        $rule = new SymbolRule('$%&@!', 0, 10);
 
         // Force generation of code coverage
-        $ruleConstruct = new Symbol('$%&@!', 0, 10);
+        $ruleConstruct = new SymbolRule('$%&@!', 0, 10);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -44,10 +44,10 @@ final class SymbolTest extends TestCase
      */
     public function testCanConstructRuleWithBothMinAndMaxConstraint(): void
     {
-        $rule = new Symbol('$%&@!', 5, 10);
+        $rule = new SymbolRule('$%&@!', 5, 10);
 
         // Force generation of code coverage
-        $ruleConstruct = new Symbol('$%&@!', 5, 10);
+        $ruleConstruct = new SymbolRule('$%&@!', 5, 10);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -58,7 +58,7 @@ final class SymbolTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        $rule = new Symbol('$%&@!', -10, null);
+        $rule = new SymbolRule('$%&@!', -10, null);
     }
 
     /**
@@ -68,7 +68,7 @@ final class SymbolTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        $rule = new Symbol('$%&@!', 10, 5);
+        $rule = new SymbolRule('$%&@!', 10, 5);
     }
 
     /**
@@ -76,10 +76,10 @@ final class SymbolTest extends TestCase
      */
     public function testCanConstructUnconstrainedRule(): void
     {
-        $rule = new Symbol('$%&@!', 0, null);
+        $rule = new SymbolRule('$%&@!', 0, null);
 
         // Force generation of code coverage
-        $ruleConstruct = new Symbol('$%&@!', 0, null);
+        $ruleConstruct = new SymbolRule('$%&@!', 0, null);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -88,10 +88,10 @@ final class SymbolTest extends TestCase
      */
     public function testCanConstructRuleWithMinConstraintEqualToMaxConstraint(): void
     {
-        $rule = new Symbol('$%&@!', 5, 5);
+        $rule = new SymbolRule('$%&@!', 5, 5);
 
         // Force generation of code coverage
-        $ruleConstruct = new Symbol('$%&@!', 5, 5);
+        $ruleConstruct = new SymbolRule('$%&@!', 5, 5);
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -102,7 +102,7 @@ final class SymbolTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        $rule = new Symbol('');
+        $rule = new SymbolRule('');
     }
 
     /**
@@ -110,7 +110,7 @@ final class SymbolTest extends TestCase
      */
     public function testCanGetCharacters(): void
     {
-        $rule = new Symbol('$%&@!');
+        $rule = new SymbolRule('$%&@!');
 
         self::assertSame('$%&@!', $rule->getCharacters());
     }
@@ -120,11 +120,11 @@ final class SymbolTest extends TestCase
      */
     public function testCanAddConstraint(): void
     {
-        $rule = new Symbol('$%&@!', 5, 5, 1);
+        $rule = new SymbolRule('$%&@!', 5, 5, 1);
         $rule->addConstraint(10, 10, 1);
 
         // Force generation of code coverage
-        $ruleConstruct = new Symbol('$%&@!', 5, 5, 1);
+        $ruleConstruct = new SymbolRule('$%&@!', 5, 5, 1);
         $ruleConstruct->addConstraint(10, 10, 1);
         self::assertEquals($rule, $ruleConstruct);
     }
@@ -134,10 +134,10 @@ final class SymbolTest extends TestCase
      */
     public function testConstraintsAreOrdered(): void
     {
-        $rule = new Symbol('$%&@!', 5, 5, 1);
+        $rule = new SymbolRule('$%&@!', 5, 5, 1);
         $rule->addConstraint(10, 10, 2);
 
-        $ruleConstruct = new Symbol('$%&@!', 10, 10, 2);
+        $ruleConstruct = new SymbolRule('$%&@!', 10, 10, 2);
         $ruleConstruct->addConstraint(5, 5, 1);
         self::assertEquals($rule, $ruleConstruct);
     }
@@ -147,7 +147,7 @@ final class SymbolTest extends TestCase
      */
     public function testMinConstraintCanBeSatisfied(): void
     {
-        $rule = new Symbol('$%&@!', 2, null);
+        $rule = new SymbolRule('$%&@!', 2, null);
 
         self::assertTrue($rule->test('FOO bar $$@'));
     }
@@ -157,7 +157,7 @@ final class SymbolTest extends TestCase
      */
     public function testMinConstraintCanBeUnsatisfied(): void
     {
-        $rule = new Symbol('$%&@!', 2, null);
+        $rule = new SymbolRule('$%&@!', 2, null);
 
         self::assertFalse($rule->test('FOO BAR $'));
     }
@@ -167,7 +167,7 @@ final class SymbolTest extends TestCase
      */
     public function testMaxConstraintCanBeSatisfied(): void
     {
-        $rule = new Symbol('$%&@!', 0, 3);
+        $rule = new SymbolRule('$%&@!', 0, 3);
 
         self::assertTrue($rule->test('FOO bar $$@'));
     }
@@ -177,7 +177,7 @@ final class SymbolTest extends TestCase
      */
     public function testMaxConstraintCanBeUnsatisfied(): void
     {
-        $rule = new Symbol('$%&@!', 0, 3);
+        $rule = new SymbolRule('$%&@!', 0, 3);
 
         self::assertFalse($rule->test('foo bar $$@!'));
     }
@@ -187,7 +187,7 @@ final class SymbolTest extends TestCase
      */
     public function testRuleIsSatisfiedWhenConstraintWeightIsLowerThanTestWeight(): void
     {
-        $rule = new Symbol('$%&@!', 0, 3, 1);
+        $rule = new SymbolRule('$%&@!', 0, 3, 1);
 
         self::assertTrue($rule->test('foo bar $$@!', 2));
     }
@@ -197,7 +197,7 @@ final class SymbolTest extends TestCase
      */
     public function testRuleCanBeValidated(): void
     {
-        $rule = new Symbol('$%&@!', 1, null);
+        $rule = new SymbolRule('$%&@!', 1, null);
 
         self::assertNull($rule->validate('&'));
     }
@@ -207,7 +207,7 @@ final class SymbolTest extends TestCase
      */
     public function testRuleWithMinConstraintCanBeInvalidated(): void
     {
-        $rule = new Symbol('$%&@!', 5, null);
+        $rule = new SymbolRule('$%&@!', 5, null);
 
         self::assertEquals(
             new ValidationError('There must be at least 5 symbols ($%&@!).', 'foo bar $$@!', $rule, 1),
@@ -220,7 +220,7 @@ final class SymbolTest extends TestCase
      */
     public function testRuleWithMaxConstraintCanBeInvalidated(): void
     {
-        $rule = new Symbol('$%&@!', 0, 10);
+        $rule = new SymbolRule('$%&@!', 0, 10);
 
         self::assertEquals(
             new ValidationError('There must be at most 10 symbols ($%&@!).', 'foo bar $$@! $$@! $$@!', $rule, 1),
@@ -233,7 +233,7 @@ final class SymbolTest extends TestCase
      */
     public function testRuleWithBothMinAndMaxConstraintCanBeInvalidated(): void
     {
-        $rule = new Symbol('$%&@!', 5, 10);
+        $rule = new SymbolRule('$%&@!', 5, 10);
 
         self::assertEquals(
             new ValidationError('There must be between 5 and 10 symbols ($%&@!).', 'foo bar $$@!', $rule, 1),
@@ -246,7 +246,7 @@ final class SymbolTest extends TestCase
      */
     public function testRuleWithMaxConstraintEqualToZeroCanBeInvalidated(): void
     {
-        $rule = new Symbol('$%&@!', 0, 0);
+        $rule = new SymbolRule('$%&@!', 0, 0);
 
         self::assertEquals(
             new ValidationError('There must be no symbols ($%&@!).', 'foo bar $$@!', $rule, 1),
@@ -259,7 +259,7 @@ final class SymbolTest extends TestCase
      */
     public function testRuleWithMinConstraintEqualToMaxConstraintCanBeInvalidated(): void
     {
-        $rule = new Symbol('$%&@!', 3, 3);
+        $rule = new SymbolRule('$%&@!', 3, 3);
 
         self::assertEquals(
             new ValidationError('There must be exactly 3 symbols ($%&@!).', 'foo bar $$@!', $rule, 1),